requirements.txt
kafka-python==2.2.15
podman-compose==1.5.0
psutil==7.1.3
py4j==0.10.9
pyspark==3.0.3
pyspark-stubs==3.0.0.post3
python-dotenv==1.2.1
PyYAML==6.0.3
